requirements.txt
Flask==2.3.6
numpy
tensorflow
Pillow
mysql-connector-python
Werkzeug
